% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hfd.R
\name{download_hfd}
\alias{download_hfd}
\title{download_hfd}
\usage{
download_hfd(url, path)
}
\arguments{
\item{url}{A string. The URL of the high-frequency data (.xlsx file).}

\item{path}{A string. The directory where the downloaded data will be stored.}
}
\value{
NULL. The function downloads the data and saves it as CSV files in the specified path.
}
\description{
Downloads high-frequency data from a specified URL and stores it in the specified path.
}
\examples{
download_hfd("https://www.ecb.europa.eu/pub/pdf/annex/Dataset_EA-MPD.xlsx", getwd())
}
