% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hfd.R
\name{load_hfd}
\alias{load_hfd}
\title{load_hfd}
\usage{
load_hfd(
  path,
  exclude_date = c("2001-09-17", "2008-10-08", "2008-11-06"),
  range = c("2001-12-31", "2018-09-13")
)
}
\arguments{
\item{path}{path to data}

\item{exclude_date}{vector of dates to exclude}

\item{range}{Defines time window. Needs to be a vector of two dates}
}
\value{

}
\description{
Load and preprocesses the Euro Area Monetary Policy Event-Study Database (EA-MPD)
}
\examples{
download_hfd("https://www.ecb.europa.eu/pub/pdf/annex/Dataset_EA-MPD.xlsx",getwd())
pcw<-load_hfd(pcw.csv,exclude_date=c("2001-09-17","2008-10-08","2008-11-06"),range=c("2001-12-31","2018-09-13"))
}
