% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings.R
\name{loadings}
\alias{loadings}
\title{loadings}
\usage{
loadings(ois_data, factors)
}
\arguments{
\item{ois_data}{A data frame containing OIS data.}

\item{factors}{A data frame containing factors.}
}
\value{
A tibble containing the loadings, standard errors, and R-squared values.
}
\description{
Computes the loadings of OIS data on factors.
}
