% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hfd.R
\name{download_hfd}
\alias{download_hfd}
\title{download_hfd}
\usage{
download_hfd(url, path)
}
\arguments{
\item{url}{url of the high frequency data}

\item{store}{Defines the download location}
}
\value{

}
\description{
Downloads the high-frequency data from a website and store it
}
\examples{
download_hfd("https://www.ecb.europa.eu/pub/pdf/annex/Dataset_EA-MPD.xlsx",getwd())
}
