% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecb_shocks.R
\name{ecb_shocks}
\alias{ecb_shocks}
\title{ecb_shocks}
\usage{
ecb_shocks(
  url = "https://www.ecb.europa.eu/pub/pdf/annex/Dataset_EA-MPD.xlsx",
  path = getwd(),
  exclude_date = c("2001-09-17", "2008-10-08", "2008-11-06"),
  range = c("2001-12-31", "2018-09-13"),
  crisis_date = "2008-09-04"
)
}
\arguments{
\item{crisis_date}{specify the starting date of the crisis for the rotation of the QE shocks}

\item{data}{specify which data to process}

\item{window}{which window do you want to rotate? choose between "release" or "conference"}
}
\value{

}
\description{
wrapper around all functions to download, process and return the ecb shocks based on Altavilla, C., Brugnolini, L., Gürkaynak, R. S., Motto, R., & Ragusa, G. (2019). Measuring euro area monetary policy. Journal of Monetary Economics, 108, 162-179.
}
\examples{
x<-ecb_shocks("https://www.ecb.europa.eu/pub/pdf/annex/Dataset_EA-MPD.xlsx",getwd(),exclude_date=c("2001-09-17","2008-10-08","2008-11-06"),range=c("2001-12-31","2018-09-13"),crisis_date="2008-09-04")
}
