% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hfd.R
\name{load_hfd}
\alias{load_hfd}
\title{load_hfd}
\usage{
load_hfd(
  data,
  exclude_dates = c("2001-09-17", "2008-10-08", "2008-11-06"),
  date_range = c("2001-12-31", "2018-09-13"),
  reproduce = FALSE,
  select_ois = TRUE
)
}
\arguments{
\item{data}{A data frame containing the raw data.}

\item{reproduce}{Logical; TRUE for exact factors Altavilla et al. 2019 factors.}

\item{select_ois}{Logical; TRUE to select OIS data.}

\item{exclude_date}{Vector of dates to exclude.}

\item{range}{Defines the time window. Needs to be a vector of two dates.}
}
\value{
A data frame containing the preprocessed data.
}
\description{
Load and preprocess the Euro Area Monetary Policy Event-Study Database (EA-MPD)
}
\examples{
data <- download_hfd("https://www.ecb.europa.eu/pub/pdf/annex/Dataset_EA-MPD.xlsx", getwd())
pcw <- load_hfd(data$pcw,
  exclude_date = c("2001-09-17", "2008-10-08", "2008-11-06"),
  range = c("2001-12-31", "2018-09-13")
)
}
